+++ index.js
@@ -1,3 +1,0 @@
-// export this package's api
-import DatePicker from './src/DatePicker';
-export default DatePicker;