+++ HISTORY.md
@@ -1,6 +1,11 @@
 # History
 ----
 
+## 6.x
+
+- Remove moment.js and change date type from Moment to Date.
+- Add use12Hours prop.
+
 ## 5.2.0 / 2016-12-17
 
 - support formatMonth/formatDay